Move tab types and /admin into tabs folder
